DraftModeParamsは二つも要らない
