+++ tekiyoke2/Assets/ScriptableObjects/DraftModeParameters.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 0}
+  m_Name: DraftModeParameters
+  m_EditorClassIdentifier: Assembly-CSharp::DraftModeParams
+  _TimeScale: 0.2
+  _DpPerSecond: 3
+  _GotDpRate: 2
