+++ tekiyoke2/Assets/ScriptableObjects/DraftModeParameters.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1168855c90a8ab949b21397c45e2f678
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
