+++ tekiyoke2/Assets/Scripts/Parameters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 23ea777b92edcc048adb6c3a2ce1c92d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
