+++ tekiyoke2/Assets/Scripts/Parameters/DraftModeParameters.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 31b7f3985b7579448815d665352556d9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
