+++ tekiyoke2/Assets/Scripts/Parameters/HeroParameters.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46fd59e3df9475144a8dd166609bb86e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
